macosx: poprawki w budowaniu, rozmiar pliku w dcc, domyślna aplikacja w ext_sound, instalacja playsound
git-svn-id: svn+ssh://kadu.net/home/kadu/repos/kadu/branches/0.4@3392 fea5205d-6df5-0310-bef7-ea67e47da9a5

+++ varia/scripts/create_macosx_bundle.sh
@@ -8,21 +8,23 @@ QTDIR=~/Desktop/root/qt
 # prefix of compiled libsndfile and openssl
 # prefiks skompilowanych bibliotek libsndfile i openssl
 SNDFILEPATH=~/Desktop/root/libsndfile
+
 #OPENSSLPATH=~/Desktop/root/openssl
 OPENSSLPATH=/usr
-
-#libao nie radzi sobie z mackiem jak trzeba
-#AOPATH=~/Desktop/root/libao
-
 # version of openssl, without letters
 # wersja openssla, bez liter
 SSLVER=0.9.7
+
+#libao nie radzi sobie z mackiem jak trzeba
+#AOPATH=~/Desktop/root/libao
 #AOVER=2.1.3
 
 INSTALLED_DIR=~/Desktop/root/kadu
 
 ICONS_PKG=ng_big_icons.tar.bz2
 ICONS_LINK=http://www.kadu.net/~joi/macosx/$ICONS_PKG
+
+PLAYSOUND_LINK=http://f.steike.com/playsound
 
 echo "Set paths in this file and remove this lines (including exit)"
 echo "Ustaw sciezki w tym pliku i usun te linie (lacznie z exit)"
@@ -47,6 +49,12 @@ if [ ! -f ${INSTALLED_DIR}/bin/kadu ]; then
 	exit
 fi
 
+if [ ! -f ${QTDIR}/include/qwidget.h ]; then
+	echo "wrong QTDIR"
+	echo "QTDIR zle ustawione"
+	exit
+fi
+
 if [ ! -f ./varia/themes/icons/default/big_message.png ]; then
 	WGETNOTFOUND=`which wget | grep "^no" | wc -l`
 	if [ ${WGETNOTFOUND} -eq 1 ]; then
@@ -62,18 +70,32 @@ if [ ! -f ./varia/themes/icons/default/big_message.png ]; then
 	cd ..
 fi
 
+if [ ! -f ./playsound ]; then
+	WGETNOTFOUND=`which wget | grep "^no" | wc -l`
+	if [ ${WGETNOTFOUND} -eq 1 ]; then
+		echo "wget not found, download this: ${PLAYSOUND_LINK} and copy it to the main directory of Kadu sources"
+		echo "wget nie znalezione, sciagnij powyzszy plik i umiesc go w glownym katalogu zrodel Kadu"
+		exit
+	fi
+	wget -c ${PLAYSOUND_LINK}
+	chmod a+x playsound
+fi
 
 KADUVERSION=`cat VERSION`
+echo "log: removing old kadu.app"
 rm -rf ${DEST}/kadu.app/
 mkdir ${DEST}/kadu.app/
 
+echo "log: copying files"
 cp -r ${INSTALLED_DIR}/share/kadu ${DEST}/kadu.app/kadu
 
 CNT_DIR=${DEST}/kadu.app/Contents
 mkdir ${CNT_DIR};
 
+echo "log: creating PkgInfo"
 echo "APPL∂çQt" > ${CNT_DIR}/PkgInfo
 
+echo "log: creating Info.plist"
 echo "<?xml version='1.0' encoding='UTF-8'?>
 <plist version='1.0'>
 <dict>
@@ -109,7 +131,12 @@ echo "<?xml version='1.0' encoding='UTF-8'?>
 
 MACOS_DIR=${CNT_DIR}/MacOS;
 mkdir ${MACOS_DIR}
+echo "log: copying kadu binary"
 cp ${INSTALLED_DIR}/bin/kadu ${MACOS_DIR}/kadu
+if [ -f ./playsound ]; then
+	echo "log: copying playsound binary"
+	cp playsound ${MACOS_DIR}/
+fi
 
 RSC_DIR=${CNT_DIR}/Resources
 mkdir ${RSC_DIR}
@@ -118,30 +145,30 @@ cp kadu/kadu.icns $RSC_DIR/
 FM_DIR=${CNT_DIR}/Frameworks
 mkdir ${FM_DIR}
 
-#if [ -f ${AOPATH}/lib/libao.2.dylib ]; then
-#	cp ${AOPATH}/lib/libao.2.dylib ${FM_DIR}
-#	cp ${AOPATH}/lib/libao.${AOVER}.dylib ${FM_DIR}
-#	mkdir -p ${FM_DIR}/ao/plugins-2
-#	cp ${AOPATH}/lib/ao/plugins-2/libmacosx.so ${FM_DIR}/ao/plugins-2/libmacosx.dylib
-#fi
+if [ -f ${OPENSSLPATH}/lib/libcrypto.${SSLVER}.dylib ]; then
+	echo "log: copying ssl libraries"
+	cp ${OPENSSLPATH}/lib/libcrypto.${SSLVER}.dylib ${FM_DIR}
+	cp ${OPENSSLPATH}/lib/libssl.${SSLVER}.dylib ${FM_DIR}
+	install_name_tool -id @executable_path/../Frameworks/libssl.${SSLVER}.dylib ${FM_DIR}/libssl.${SSLVER}.dylib
+	install_name_tool -change ${OPENSSLPATH}/lib/libcrypto.${SSLVER}.dylib @executable_path/../Frameworks/libcrypto.${SSLVER}.dylib ${FM_DIR}/libssl.${SSLVER}.dylib
+	install_name_tool -id @executable_path/../Frameworks/libcrypto.${SSLVER}.dylib ${FM_DIR}/libcrypto.${SSLVER}.dylib
+fi
 
-cp ${SNDFILEPATH}/lib/libsndfile.1.dylib ${FM_DIR}
-cp ${OPENSSLPATH}/lib/libcrypto.${SSLVER}.dylib ${OPENSSLPATH}/lib/libssl.${SSLVER}.dylib ${FM_DIR}
-cp ${QTDIR}/lib/libqt-mt.3.dylib ${FM_DIR}
+if [ -f ${QTDIR}/lib/libqt-mt.3.dylib ]; then
+	echo "log: copying qt library"
+	cp ${QTDIR}/lib/libqt-mt.3.dylib ${FM_DIR}
+	install_name_tool -id @executable_path/../Frameworks/libqt-mt.3.dylib ${FM_DIR}/libqt-mt.3.dylib
+fi
 
-install_name_tool -id @executable_path/../Frameworks/libqt-mt.3.dylib ${FM_DIR}/libqt-mt.3.dylib
-install_name_tool -id @executable_path/../Frameworks/libsndfile.1.dylib ${FM_DIR}/libsndfile.1.dylib
-install_name_tool -id @executable_path/../Frameworks/libssl.${SSLVER}.dylib ${FM_DIR}/libssl.${SSLVER}.dylib
-#install_name_tool -id @executable_path/../Frameworks/libcrypto.${SSLVER}.dylib ${FM_DIR}/libssl.${SSLVER}.dylib
-install_name_tool -change ${OPENSSLPATH}/lib/libcrypto.${SSLVER}.dylib @executable_path/../Frameworks/libcrypto.${SSLVER}.dylib ${FM_DIR}/libssl.${SSLVER}.dylib
-install_name_tool -id @executable_path/../Frameworks/libcrypto.${SSLVER}.dylib ${FM_DIR}/libcrypto.${SSLVER}.dylib
-#if [ -f ${DEST}/kadu.app/kadu/modules/ao_sound.dylib ]; then
-#	install_name_tool -id @executable_path/../Frameworks/libao.2.dylib ${FM_DIR}/libao.2.dylib
-#fi
+if [ -f ${SNDFILEPATH}/lib/libsndfile.1.dylib ]; then
+	echo "log: copying sndfile library"
+	cp ${SNDFILEPATH}/lib/libsndfile.1.dylib ${FM_DIR}
+	install_name_tool -id @executable_path/../Frameworks/libsndfile.1.dylib ${FM_DIR}/libsndfile.1.dylib
+fi
 
 cd ${MACOS_DIR}
+echo "log: changing library bindings"
 install_name_tool -change libqt-mt.3.dylib @executable_path/../Frameworks/libqt-mt.3.dylib ./kadu
-
 install_name_tool -change ${SNDFILEPATH}/lib/libsndfile.1.dylib @executable_path/../Frameworks/libsndfile.1.dylib ./kadu
 install_name_tool -change ${OPENSSLPATH}/lib/libssl.${SSLVER}.dylib @executable_path/../Frameworks/libssl.${SSLVER}.dylib ./kadu
 install_name_tool -change ${OPENSSLPATH}/lib/libcrypto.${SSLVER}.dylib @executable_path/../Frameworks/libcrypto.${SSLVER}.dylib ./kadu
@@ -153,8 +180,18 @@ if [ -f ${DEST}/kadu.app/kadu/modules/encryption.dylib ]; then
 	install_name_tool -change ${OPENSSLPATH}/lib/libssl.${SSLVER}.dylib @executable_path/../Frameworks/libssl.${SSLVER}.dylib ${DEST}/kadu.app/kadu/modules/encryption.dylib
 	install_name_tool -change ${OPENSSLPATH}/lib/libcrypto.${SSLVER}.dylib @executable_path/../Frameworks/libcrypto.${SSLVER}.dylib ${DEST}/kadu.app/kadu/modules/encryption.dylib
 fi
+
+#if [ -f ${AOPATH}/lib/libao.2.dylib ]; then
+#	cp ${AOPATH}/lib/libao.2.dylib ${FM_DIR}
+#	cp ${AOPATH}/lib/libao.${AOVER}.dylib ${FM_DIR}
+#	mkdir -p ${FM_DIR}/ao/plugins-2
+#	cp ${AOPATH}/lib/ao/plugins-2/libmacosx.so ${FM_DIR}/ao/plugins-2/libmacosx.dylib
+#fi
+#if [ -f ${DEST}/kadu.app/kadu/modules/ao_sound.dylib ]; then
+#	install_name_tool -id @executable_path/../Frameworks/libao.2.dylib ${FM_DIR}/libao.2.dylib
+#fi
 #if [ -f ${DEST}/kadu.app/kadu/modules/ao_sound.dylib ]; then
 #	install_name_tool -change ${AOPATH}/lib/libao.2.dylib @executable_path/../Frameworks/libao.2.dylib ${DEST}/kadu.app/kadu/modules/ao_sound.dylib
 #fi
 
-echo "done"
+echo "everything done"
